+++ interfaces/Portal/src/app/services/programs-service-api.service.ts
@@ -4,6 +4,7 @@ import { saveAs } from 'file-saver';
 import { environment } from '../../environments/environment';
 import { CURRENT_USER_ENDPOINT_PATH } from '../auth/auth.service';
 import { UserRole } from '../auth/user-role.enum';
+import { FilterOperator, FilterParameter } from '../enums/filters.enum';
 import RegistrationStatus from '../enums/registration-status.enum';
 import { ActionType, LatestAction } from '../models/actions.model';
 import { Event } from '../models/event.model';
@@ -30,21 +31,13 @@ import { Role, TableData, User, UserSearchResult } from '../models/user.model';
 import { ImportResult } from '../program/bulk-import/bulk-import.component';
 import { arrayToXlsx } from '../shared/array-to-xlsx';
 import { ApiService } from './api.service';
-import {
-  FilterOperatorEnum,
-  FilterService,
-  PaginationFilter,
-  PaginationSort,
-} from './filter.service';
+import { PaginationFilter, PaginationSort } from './filter.service';
 
 @Injectable({
   providedIn: 'root',
 })
 export class ProgramsServiceApiService {
-  constructor(
-    private apiService: ApiService,
-    private filterService: FilterService,
-  ) {}
+  constructor(private apiService: ApiService) {}
 
   login(username: string, password: string): Promise<User | null> {
     console.log('ProgramsService : login()');
@@ -97,7 +90,7 @@ export class ProgramsServiceApiService {
     dryRun = false,
     filters?: PaginationFilter[],
   ): Promise<any> {
-    const params = this.filterToParams(filters, dryRun);
+    const params = this.filtersToParams(filters, dryRun);
     return this.apiService.delete(
       environment.url_121_service_api,
       `/programs/${programId}/registrations`,
@@ -299,7 +292,7 @@ export class ProgramsServiceApiService {
     dryRun = false,
     filters?: PaginationFilter[],
   ): Promise<any> {
-    const params = this.filterToParams(filters, dryRun);
+    const params = this.filtersToParams(filters, dryRun);
     return this.apiService.post(
       environment.url_121_service_api,
       `/programs/${programId}/payments`,
@@ -335,7 +328,7 @@ export class ProgramsServiceApiService {
   ): Promise<void> {
     const downloadData: string[] = await this.apiService.get(
       environment.url_121_service_api,
-      `/programs/${programId}/registrations/import-template/${type}`,
+      `/programs/${programId}/registrations/import-template`,
       false,
     );
 
@@ -476,14 +469,11 @@ export class ProgramsServiceApiService {
         );
       }
       if (allPeopleAffectedOptions.filters) {
-        for (const filter of allPeopleAffectedOptions.filters) {
-          const defaultFilter = FilterOperatorEnum.ilike;
-          const operator = filter.operator ? filter.operator : defaultFilter;
-          params = params.append(
-            `filter.${filter.name}`,
-            `${operator}:${filter.value}`,
-          );
-        }
+        params = this.filtersToParams(
+          allPeopleAffectedOptions?.filters,
+          false,
+          params,
+        );
       }
       if (allPeopleAffectedOptions.sort) {
         params = params.append(
@@ -602,18 +592,22 @@ export class ProgramsServiceApiService {
     statuses?: RegistrationStatus[],
     filters?: PaginationFilter[],
     sort?: PaginationSort,
-    // TODO: Fix the 'any' for the 'links' parameter
-  ): Promise<{ data: Person[]; meta: PaginationMetadata; links: any }> {
-    const quickSearch = filters
-      ? filters.find(
-          (filter) =>
-            filter.name === this.filterService.DEFAULT_FILTER_OPTION.name,
-        )
-      : null;
-
+  ): Promise<{
+    data: Person[];
+    meta: PaginationMetadata;
+    links: {
+      current: string;
+      first: string;
+      last: string;
+      next: string;
+      previous: string;
+    };
+  }> {
     let params = new HttpParams();
+
     params = params.append('limit', limit);
     params = params.append('page', page);
+
     // TODO: This still needs to be added to the back-end in a future item
     if (referenceId) {
       params = params.append('filter.referenceId', referenceId);
@@ -627,26 +621,16 @@ export class ProgramsServiceApiService {
     if (statuses) {
       params = params.append(
         'filter.status',
-        `${FilterOperatorEnum.in}:${statuses.join(',')}`,
+        `${FilterOperator.in}:${statuses.join(',')}`,
       );
     }
-    if (quickSearch) {
-      params = params.append('search', quickSearch.value);
-      filters = filters.filter((filter) => filter.name !== quickSearch.name);
-    }
     if (filters) {
-      for (const filter of filters) {
-        const defaultFilter = FilterOperatorEnum.ilike;
-        const operator = filter.operator ? filter.operator : defaultFilter;
-        params = params.append(
-          `filter.${filter.name}`,
-          `${operator}:${filter.value}`,
-        );
-      }
+      params = this.filtersToParams(filters, false, params);
     }
     if (sort) {
       params = params.append('sortBy', `${sort.column}:${sort.direction}`);
     }
+
     const { data, meta, links } = await this.apiService.get(
       environment.url_121_service_api,
       `/programs/${programId}/registrations`,
@@ -665,7 +649,7 @@ export class ProgramsServiceApiService {
     message?: string,
     messageTemplateKey?: string,
   ): Promise<any> {
-    const params = this.filterToParams(filters, dryRun);
+    const params = this.filtersToParams(filters, dryRun);
     return this.apiService.patch(
       environment.url_121_service_api,
       `/programs/${programId}/registrations/status`,
@@ -782,7 +766,7 @@ export class ProgramsServiceApiService {
     filters?: PaginationFilter[],
     messageTemplateKey?: string,
   ): Promise<any> {
-    const params = this.filterToParams(filters, dryRun);
+    const params = this.filtersToParams(filters, dryRun);
     return this.apiService.post(
       environment.url_121_service_api,
       `/programs/${programId}/registrations/message`,
@@ -987,15 +971,29 @@ export class ProgramsServiceApiService {
       });
   }
 
-  private filterToParams(
+  private filtersToParams(
     filters: PaginationFilter[],
     dryRun: boolean,
+    existingParams?: HttpParams,
   ): HttpParams {
-    let params = new HttpParams();
+    let params: HttpParams;
+
+    if (existingParams) {
+      params = existingParams;
+    } else {
+      params = new HttpParams();
+    }
+
     params = params.append('dryRun', dryRun);
+
     if (filters) {
       for (const filter of filters) {
-        const defaultFilter = FilterOperatorEnum.ilike;
+        if (filter.name === FilterParameter.search) {
+          params = params.append(FilterParameter.search, filter.value);
+          continue;
+        }
+
+        const defaultFilter = FilterOperator.ilike;
         const operator = filter.operator ? filter.operator : defaultFilter;
         params = params.append(
           `filter.${filter.name}`,
@@ -1003,6 +1001,7 @@ export class ProgramsServiceApiService {
         );
       }
     }
+
     return params;
   }