+++ services/121-service/src/registration/services/registrations-pagination.service.ts
@@ -15,10 +15,9 @@ import {
   WhereExpressionBuilder,
 } from 'typeorm';
 import { FspName } from '../../fsp/enum/fsp-name.enum';
-import { FinancialServiceProviderEntity } from '../../fsp/financial-service-provider.entity';
-import { ProgramFspConfigurationService } from '../../programs/fsp-configuration/fsp-configuration.service';
 import { ProgramEntity } from '../../programs/program.entity';
 import { ProgramService } from '../../programs/programs.service';
+import { overwriteProgramFspDisplayName } from '../../programs/utils/overwrite-fsp-display-name.helper';
 import { ScopedQueryBuilder } from '../../scoped.repository';
 import { StatusEnum } from '../../shared/enum/status.enum';
 import { PermissionEnum } from '../../user/enum/permission.enum';
@@ -51,13 +50,10 @@ export class RegistrationsPaginationService {
   private readonly programRepository: Repository<ProgramEntity>;
   @InjectRepository(UserEntity)
   private readonly userRepository: Repository<UserEntity>;
-  @InjectRepository(FinancialServiceProviderEntity)
-  private readonly financialServiceProviderRepository: Repository<FinancialServiceProviderEntity>;
 
   public constructor(
     private readonly programService: ProgramService,
     private readonly registrationViewScopedRepository: RegistrationViewScopedRepository,
-    private readonly programFspConfigurationService: ProgramFspConfigurationService,
   ) {}
 
   public async getPaginate(
@@ -67,6 +63,9 @@ export class RegistrationsPaginationService {
     noLimit: boolean,
     queryBuilder?: ScopedQueryBuilder<RegistrationViewEntity>,
   ): Promise<Paginated<RegistrationViewEntity>> {
+    // Deep clone query here to prevent mutation out of this function
+    query = structuredClone(query);
+
     let paginateConfigCopy = { ...PaginateConfigRegistrationView };
     if (noLimit) {
       // These setting are needed to get all registrations
@@ -82,6 +81,13 @@ export class RegistrationsPaginationService {
     if (query.select && query.select.includes('name')) {
       if (fullnameNamingConvention) {
         query.select = query.select.concat(fullnameNamingConvention);
+      }
+    }
+
+    // If you want to select fspDisplayName, you also need to get financialServiceProvider because we need this to find the correct fspDisplayName
+    if (query.select && query.select.includes('fspDisplayName')) {
+      if (fullnameNamingConvention) {
+        query.select.push('financialServiceProvider');
       }
     }
 
@@ -118,6 +124,11 @@ export class RegistrationsPaginationService {
       );
     }
 
+    if (query.search) {
+      queryBuilder = this.addSearchToQueryBuilder(queryBuilder, query.search);
+      delete query.search;
+    }
+
     // Check if the sort contains at least one registration data name
     // At the moment we only support sorting on one field
     if (
@@ -133,7 +144,7 @@ export class RegistrationsPaginationService {
     }
 
     if (hasPersonalReadPermission) {
-      paginateConfigCopy.relations = ['data', 'dataSearchBy'];
+      paginateConfigCopy.relations = ['data'];
     } else {
       paginateConfigCopy.searchableColumns = [];
     }
@@ -163,6 +174,7 @@ export class RegistrationsPaginationService {
       orignalSelect,
       fullnameNamingConvention,
       hasPersonalReadPermission,
+      programId,
     );
     return result;
   }
@@ -209,6 +221,17 @@ export class RegistrationsPaginationService {
       programId,
       paginateQuery,
     );
+  }
+
+  private addSearchToQueryBuilder(
+    queryBuilder: ScopedQueryBuilder<RegistrationViewEntity>,
+    search: string,
+  ): ScopedQueryBuilder<RegistrationViewEntity> {
+    queryBuilder.leftJoin('registration.data', 'registrationDataSearch');
+    queryBuilder.andWhere('registrationDataSearch.value ILIKE :search', {
+      search: `%${search}%`,
+    });
+    return queryBuilder;
   }
 
   private async throwIfNoTransactionReadPermission(
@@ -442,7 +465,10 @@ export class RegistrationsPaginationService {
     orignalSelect: string[],
     fullnameNamingConvention: string[],
     hasPersonalReadPermission: boolean,
+    programId: number,
   ): Promise<RegistrationViewEntity[]> {
+    const fspDisplayNameMapping =
+      await this.getFspDisplayNameMapping(programId);
     const mappedData: RegistrationViewEntity[] = [];
     for (const registration of paginatedResult.data) {
       const mappedRootRegistration = this.mapRootRegistration(
@@ -456,8 +482,19 @@ export class RegistrationsPaginationService {
         mappedRootRegistration,
         registrationDataRelations,
       );
-      mappedRegistration.fspDisplayName =
-        await this.overwriteFspDisplayName(mappedRegistration);
+      if (!orignalSelect || orignalSelect?.includes('fspDisplayName')) {
+        mappedRegistration.fspDisplayName = await this.overwriteFspDisplayName(
+          mappedRegistration,
+          fspDisplayNameMapping,
+        );
+        if (
+          orignalSelect &&
+          !orignalSelect.includes('financialServiceProvider')
+        ) {
+          delete mappedRegistration.financialServiceProvider;
+        }
+      }
+
       mappedData.push(mappedRegistration);
 
       if ((!select || select.includes('name')) && hasPersonalReadPermission) {
@@ -472,33 +509,33 @@ export class RegistrationsPaginationService {
     return mappedData;
   }
 
+  private async getFspDisplayNameMapping(
+    programId: number,
+  ): Promise<Record<string, JSON>> {
+    const map = {};
+    const program = await this.programRepository.findOne({
+      where: { id: programId },
+      relations: ['financialServiceProviders', 'programFspConfiguration'],
+    });
+    if (program.financialServiceProviders.length > 0) {
+      program.financialServiceProviders = overwriteProgramFspDisplayName(
+        program.financialServiceProviders,
+        program.programFspConfiguration,
+      );
+    }
+    for (const fsp of program.financialServiceProviders) {
+      map[fsp.fsp] = fsp.displayName;
+    }
+    return map;
+  }
+
   private async overwriteFspDisplayName(
     registration: RegistrationViewEntity,
-  ): Promise<string> {
+    fspDisplayNameMapping: Record<string, JSON>,
+  ): Promise<JSON> {
     if (registration.financialServiceProvider) {
-      const financialServiceProvider =
-        await this.financialServiceProviderRepository.findOne({
-          where: { fsp: registration.financialServiceProvider },
-        });
-
-      if (financialServiceProvider) {
-        const programFspDisplayNameConfig =
-          await this.programFspConfigurationService.findDisplayNameConfiguration(
-            registration.programId,
-            financialServiceProvider.id,
-          );
-
-        if (programFspDisplayNameConfig) {
-          registration.fspDisplayName = Object.assign(
-            {},
-            financialServiceProvider.displayName,
-            programFspDisplayNameConfig.value,
-          ) as string;
-        }
-      }
+      return fspDisplayNameMapping[registration.financialServiceProvider];
     }
-
-    return registration.fspDisplayName;
   }
 
   private mapRootRegistration(
@@ -518,7 +555,6 @@ export class RegistrationsPaginationService {
       mappedRegistration = { ...registration };
     }
     delete mappedRegistration.data;
-    delete mappedRegistration.dataSearchBy;
     if (!hasPersonalReadPermission) {
       delete mappedRegistration.phoneNumber;
     }
@@ -655,17 +691,19 @@ export class RegistrationsPaginationService {
     status: StatusEnum,
     paymentNumber: string,
   ): ScopedQueryBuilder<RegistrationViewEntity> {
+    const paymentNumberKey = `${alias}PaymentNumber`;
+    const statusKey = `${alias}Status`;
     queryBuilder.leftJoin(
       'registration.latestTransactions',
       alias,
-      `"${alias}"."payment" = :paymentNumber`,
-      { paymentNumber: paymentNumber },
+      `"${alias}"."payment" = :${paymentNumberKey}`,
+      { [paymentNumberKey]: paymentNumber },
     );
     if (status) {
       queryBuilder
         .innerJoin(`${alias}.transaction`, `transaction${alias}`)
-        .andWhere(`"transaction${alias}"."status" = :status`, {
-          status: status,
+        .andWhere(`"transaction${alias}"."status" = :${statusKey}`, {
+          [statusKey]: status,
         });
     } else {
       queryBuilder.andWhere(`"${alias}"."id" IS NULL`);