+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/global-121/121-platform/compare/v1.116.0...main)
+## [Unreleased](https://github.com/global-121/121-platform/compare/v1.116.9...main)
 
 ### Removed
 
@@ -13,6 +13,48 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - AW-App interface code and all related references
 
 ---
+
+## [1.116.10](https://github.com/global-121/121-platform/compare/v1.116.9...v1.116.10)- 2024-04-24
+
+### Fixed
+
+- Export registrations fix fsp displayname and make it default always english
+
+## [1.116.9](https://github.com/global-121/121-platform/compare/v1.116.8...v1.116.9)- 2024-04-22
+
+### Fixed
+
+- Handling of FSP's "Display names" in the "Edit PA"-popup
+
+## [1.116.8](https://github.com/global-121/121-platform/compare/v1.116.7...v1.116.8)- 2024-04-19
+
+### Disabled
+
+- Quick search in frontend
+
+## [1.116.7](https://github.com/global-121/121-platform/compare/v1.116.6...v1.116.7)- 2024-04-18
+
+### Fixed
+
+- Performance issues for PV program
+- Heap out of memory when exporting
+- Quick search not applied when using 'Export filtered view'
+
+## [1.116.6](https://github.com/global-121/121-platform/compare/v1.116.5...v1.116.6)- 2024-04-16
+
+## [1.116.5](https://github.com/global-121/121-platform/compare/v1.116.4...v1.116.5)- 2024-04-15
+
+## [1.116.4](https://github.com/global-121/121-platform/compare/v1.116.3...v1.116.4)- 2024-04-15
+
+### Fixed
+
+- Calculate payment amount memory error
+
+## [1.116.3](https://github.com/global-121/121-platform/compare/v1.116.2...v1.116.3)- 2024-04-15
+
+### Fixed
+
+- Export in smaller chunks to prevent memory error
 
 ## [1.116.2](https://github.com/global-121/121-platform/compare/v1.116.1...v1.116.2)- 2024-04-10